feat: centralize image placeholder utilities
