+++ components/cover-image.tsx
@@ -2,7 +2,7 @@ import cn from "classnames";
 import Image from "next/image";
 import Link from "next/link";
 
-import { placeholderImage } from "./intro";
+import { placeholderImage } from "../lib/image";
 
 type Props = {
   title: string;
